deep_search_engine/requirements.txt
# Deep Learning Framework
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# NLP Libraries
spacy>=3.6.0
nltk>=3.8.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0